+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# Changelog for Cldr v3.2.5
+
+This is the changelog for Cldr v2.2.6 released on ______, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr/tags)
+
+### Enhancements
+
+* Adds the option `:round_up_to` to `Money.round/2`. This will set the fractional part of a money to the specified integer amount.  As examples:
+```
+  Money.round Money.new("123.76", :USD), round_up_to: 99
+  #Money<:USD, 123.99>
+
+  # JPY has no fractional digits so this is equivalent to
+  # rounding to the next yen
+  Money.round Money.new("123.7456", :JPY), round_up_to: 99
+  #Money<:JPY, 124>
+```
+
+### Bug Fixes
+
+* Parsing money strings now uses a more complete set of character definitions for decimal and grouping separators based upon the `characters.json` file of the "en" locale.
+
 # Changelog for Money v3.2.4
 
 This is the changelog for Money v3.2.4 released on February 13th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/money/tags)